third_party/blink/web_tests/storage/indexeddb/interfaces-expected.txt

Test IndexedDB's interfaces.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS 'IDBCursor' in self is true
PASS 'IDBCursorWithValue' in self is true
PASS 'IDBDatabase' in self is true
PASS 'IDBFactory' in self is true
PASS 'IDBIndex' in self is true
PASS 'IDBKeyRange' in self is true
PASS 'IDBObjectStore' in self is true
PASS 'IDBOpenDBRequest' in self is true
PASS 'IDBRequest' in self is true
PASS 'IDBTransaction' in self is true
PASS 'IDBVersionChangeEvent' in self is true
PASS successfullyParsed is true

TEST COMPLETE
